Remove faq from workspaces detail page
